Reworked responsive sidebar ui: moved meaningful content into server component for prerendering with proper html structuring
