+++ packages_dependencies.txt
@@ -1,0 +1,3 @@
+js-cookie
+lucide-react
+next-themes
