+++ packages_devDependencies.txt
@@ -1,0 +1,1 @@
+@types/js-cookie@3.0.6
